Use `StringInput` for all text inputs
