+++ panel/src/components/Forms/Input/EmailInput.vue
@@ -1,8 +1,18 @@
+<template>
+	<k-string-input
+		v-bind="$props"
+		type="email"
+		class="k-email-input"
+		@input="$emit('input', $event)"
+		@invalid="($invalid, $v) => $emit('invalid', $invalid, $v)"
+	/>
+</template>
+
 <script>
-import TextInput, { props as TextInputProps } from "./TextInput.vue";
+import StringInput, { props as StringInputProps } from "./StringInput.vue";
 
 export const props = {
-	mixins: [TextInputProps],
+	mixins: [StringInputProps],
 	props: {
 		autocomplete: {
 			type: String,
@@ -11,10 +21,6 @@ export const props = {
 		placeholder: {
 			type: String,
 			default: () => window.panel.$t("email.placeholder")
-		},
-		type: {
-			type: String,
-			default: "email"
 		}
 	}
 };
@@ -23,7 +29,6 @@ export const props = {
  * @example <k-email-input :value="email" @input="email = $event" name="email" />
  */
 export default {
-	extends: TextInput,
-	mixins: [props]
+	mixins: [StringInput]
 };
 </script>
